Update dtype for strides to int64 to accomodate runtime changes

+++ pjrt_implementation/src/api/loaded_executable_instance.cc
@@ -261,7 +261,7 @@ tt_pjrt_status LoadedExecutableInstance::createDefaultOutputBuffers(
 
     // Row-major strides: last stride is 1, each preceding stride is the
     // product of all following dimension sizes.
-    std::vector<std::uint32_t> strides(output_shape->size());
+    std::vector<std::int64_t> strides(output_shape->size());
     std::exclusive_scan(output_shape->rbegin(), output_shape->rend(),
                         strides.rbegin(), std::uint32_t(1),
                         std::multiplies<>());
